+++ src/Spec/VocaLinkV590.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Cs278\BankModulus\Spec;
+
+use Cs278\BankModulus\BankAccountNormalized;
+use Cs278\BankModulus\Spec\VocaLinkV380\DataV590;
+use Cs278\BankModulus\Spec\VocaLinkV380\Driver;
+
+final class VocaLinkV590 implements SpecInterface
+{
+    /** @var Driver */
+    private $driver;
+
+    public function __construct()
+    {
+        $this->driver = new Driver(new DataV590());
+    }
+
+    public function check(BankAccountNormalized $bankAccount)
+    {
+        return $this->driver->check($bankAccount);
+    }
+}